+++ src/engine/Component/Camera/CameraComponent.cpp
@@ -57,7 +57,7 @@ namespace vg::engine
     }
 
     //--------------------------------------------------------------------------------------
-    void CameraComponent::Update(double _dt)
+    void CameraComponent::Update(float _dt)
     {
         auto * view = Engine::get()->GetRenderer()->GetView(gfx::ViewID(m_target, m_index));
         if (nullptr != view)
